build: make the Windows and macOS builds actually possible
Three gaps would have broken the cross-platform jobs the moment they ran:

- No vcpkg manifest. The Windows job invoked vcpkg but nothing told it what to
  install, so find_package(SQLite3) would have failed immediately. Added
  vcpkg.json with the dependency set and a baseline pinned to vcpkg 2026.07.29,
  and pinned the workflows' vcpkg checkout to the same commit so the toolchain
  and the manifest cannot disagree.
- TagLib was found only through pkg_check_modules. pkg-config is generally
  absent on Windows and vcpkg ships a CMake config package instead, so the
  Windows build could not have located it. Discovery now tries the config
  package first and falls back to pkg-config, exposing one Resonance::TagLib
  target either way. On this host it now takes the config route.
- No CMakePresets.json, which FRD section 3 asks for. Added presets for all
  four targets plus strict and sanitiser variants, and the workflows now use
  them so CI cannot drift from what a developer runs locally.

Fixed five -Werror=conversion failures in the desktop sources. The CI Linux job
builds with warnings as errors and would have failed on all five; verified with
a clean -Werror build and the full suite.

docs/building.md covers both routes and states plainly that only Linux x64 has
actually been built.

+++ src/MusicLibrary.Desktop/src/Workbench.cpp
@@ -563,7 +563,8 @@ void DeckPanel::setPlaybackAvailable(bool available, const QString& deviceName) 
 void DeckPanel::setPosition(double fraction) {
 	m_position = fraction;
 	if (!m_record) return;
-	const std::int64_t atMs = static_cast<std::int64_t>(fraction * m_record->audio.durationMs);
+	const std::int64_t atMs = static_cast<std::int64_t>(
+		fraction * static_cast<double>(m_record->audio.durationMs));
 	m_counterLabel->setText(qs(text::formatDuration(atMs)) + QStringLiteral(" / ")
 		+ qs(text::formatDuration(m_record->audio.durationMs)));
 }
@@ -576,7 +577,7 @@ void DeckPanel::updateMarquee() {
 		m_marqueeLabel->setText(m_marqueeText);
 		return;
 	}
-	m_marqueeOffset = (m_marqueeOffset + 1) % m_marqueeText.size();
+	m_marqueeOffset = (m_marqueeOffset + 1) % static_cast<int>(m_marqueeText.size());
 	const QString doubled = m_marqueeText + m_marqueeText;
 	m_marqueeLabel->setText(doubled.mid(m_marqueeOffset, kWidth));
 }
